[Decomp][Group] Cohesion split: extract member-state + instances/rewards to Group*.cpp
Mirror of mangosthree #266. Pure cohesion split of Group.cpp with no
behaviour change:

- GroupInstance.cpp     - instance binding/reset, dungeon/raid difficulty,
  battleground-queue eligibility, looter rotation, group kill rewards
  (incl. the static RewardGroupAtKill_helper).
- GroupMemberState.cpp  - low-level roster mutation: add/remove, leader/
  assistant/main-tank flags, sub-group assignment, roll cleanup.

Group.h is unchanged. CMake file(GLOB) auto-compiles the new TUs. WotLK
has no _updateLeaderFlag (Cata-only), so it stays a 13/13 split here.
