+++ src/game/WorldHandlers/Group.cpp
@@ -1593,1157 +1593,29 @@ void Group::OfflineReadyCheck()
     }
 }
 
-/**
- * @brief Adds a member to the first available subgroup.
- *
- * @param guid The member player GUID.
- * @param name The member player name.
- * @param isAssistant True if the member should be marked as assistant.
- * @return true if a suitable subgroup was found and the member was added; otherwise false.
- */
-bool Group::_addMember(ObjectGuid guid, const char* name, bool isAssistant)
-{
-    // get first not-full group
-    uint8 groupid = 0;
-    if (m_subGroupsCounts)
-    {
-        bool groupFound = false;
-        for (; groupid < MAX_RAID_SUBGROUPS; ++groupid)
-        {
-            if (m_subGroupsCounts[groupid] < MAX_GROUP_SIZE)
-            {
-                groupFound = true;
-                break;
-            }
-        }
-        // We are raid group and no one slot is free
-        if (!groupFound)
-        {
-            return false;
-        }
-    }
 
-    return _addMember(guid, name, isAssistant, groupid);
-}
 
-/**
- * @brief Adds a member to a specific subgroup and updates player references.
- *
- * @param guid The member player GUID.
- * @param name The member player name.
- * @param isAssistant True if the member should be marked as assistant.
- * @param group The subgroup assignment.
- * @return true if the member was added; otherwise false.
- */
-bool Group::_addMember(ObjectGuid guid, const char* name, bool isAssistant, uint8 group)
-{
-    if (IsFull())
-    {
-        return false;
-    }
 
-    if (!guid)
-    {
-        return false;
-    }
 
-    Player* player = sObjectMgr.GetPlayer(guid, false);
 
-    uint32 lastMap = 0;
-    if (player && player->IsInWorld())
-    {
-        lastMap = player->GetMapId();
-    }
-    else if (player && player->IsBeingTeleported())
-    {
-        lastMap = player->GetTeleportDest().mapid;
-    }
 
-    MemberSlot member;
-    member.guid      = guid;
-    member.name      = name;
-    member.group     = group;
-    member.assistant = isAssistant;
-    member.lastMap   = lastMap;
-    m_memberSlots.push_back(member);
 
-    SubGroupCounterIncrease(group);
 
-    if (player)
-    {
-        player->SetGroupInvite(NULL);
-        // if player is in group and he is being added to BG raid group, then call SetBattleGroundRaid()
-        if (player->GetGroup() && isBGGroup())
-        {
-            player->SetBattleGroundRaid(this, group);
-        }
-        // if player is in bg raid and we are adding him to normal group, then call SetOriginalGroup()
-        else if (player->GetGroup())
-        {
-            player->SetOriginalGroup(this, group);
-        }
-        // if player is not in group, then call set group
-        else
-        {
-            player->SetGroup(this, group);
-        }
 
-        if (player->IsInWorld())
-        {
-            // if the same group invites the player back, cancel the homebind timer
-            if (InstanceGroupBind* bind = GetBoundInstance(player->GetMapId(), player))
-                if (bind->state->GetInstanceId() == player->GetInstanceId())
-                {
-                    player->m_InstanceValid = true;
-                }
-        }
-    }
 
-    if (!isRaidGroup())                                     // reset targetIcons for non-raid-groups
-    {
-        for (int i = 0; i < TARGET_ICON_COUNT; ++i)
-        {
-            m_targetIcons[i].Clear();
-        }
-    }
 
-    if (!isBGGroup())
-    {
-        // insert into group table
-        CharacterDatabase.PExecute("INSERT INTO `group_member` (`groupId`,`memberGuid`,`assistant`,`subgroup`) VALUES ('%u','%u','%u','%u')",
-                                   m_Id, member.guid.GetCounter(), ((member.assistant == 1) ? 1 : 0), member.group);
-    }
 
-    return true;
-}
 
-/**
- * @brief Removes a member from internal group state and updates leadership if needed.
- *
- * @param guid The member player GUID.
- * @return true if the removed member was the leader and leadership changed; otherwise false.
- */
-bool Group::_removeMember(ObjectGuid guid)
-{
-    Player* player = sObjectMgr.GetPlayer(guid);
-    if (player)
-    {
-        // if we are removing player from battleground raid
-        if (isBGGroup())
-        {
-            player->RemoveFromBattleGroundRaid();
-        }
-        else
-        {
-            // we can remove player who is in battleground from his original group
-            if (player->GetOriginalGroup() == this)
-            {
-                player->SetOriginalGroup(NULL);
-            }
-            else
-            {
-                player->SetGroup(NULL);
-            }
-        }
-    }
 
-    _removeRolls(guid);
 
-    member_witerator slot = _getMemberWSlot(guid);
-    if (slot != m_memberSlots.end())
-    {
-        SubGroupCounterDecrease(slot->group);
 
-        m_memberSlots.erase(slot);
-    }
 
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("DELETE FROM `group_member` WHERE `memberGuid`='%u'", guid.GetCounter());
-    }
 
-    if (m_leaderGuid == guid)                               // leader was removed
-    {
-        if (GetMembersCount() > 0)
-        {
-            _setLeader(m_memberSlots.front().guid);
-        }
-        return true;
-    }
 
-    return false;
-}
 
-/**
- * @brief Updates the stored group leader and migrates instance bindings as needed.
- *
- * @param guid The new leader GUID.
- */
-void Group::_setLeader(ObjectGuid guid)
-{
-    member_citerator slot = _getMemberCSlot(guid);
-    if (slot == m_memberSlots.end())
-    {
-        return;
-    }
 
-    if (!isBGGroup())
-    {
-        uint32 slot_lowguid = slot->guid.GetCounter();
 
-        uint32 leader_lowguid = m_leaderGuid.GetCounter();
 
-        // TODO: set a time limit to have this function run rarely cause it can be slow
-        CharacterDatabase.BeginTransaction();
 
-        // update the group's bound instances when changing leaders
 
-        // remove all permanent binds from the group
-        // in the DB also remove solo binds that will be replaced with permbinds
-        // from the new leader
-        CharacterDatabase.PExecute(
-            "DELETE FROM `group_instance` WHERE `leaderguid`='%u' AND (`permanent` = 1 OR "
-            "`instance` IN (SELECT `instance` FROM `character_instance` WHERE `guid` = '%u')"
-            ")", leader_lowguid, slot_lowguid);
 
-        Player* player = sObjectMgr.GetPlayer(slot->guid);
-
-        if (player)
-        {
-            for (uint8 i = 0; i < MAX_DIFFICULTY; ++i)
-            {
-                for (BoundInstancesMap::iterator itr = m_boundInstances[i].begin(); itr != m_boundInstances[i].end();)
-                {
-                    if (itr->second.perm)
-                    {
-                        itr->second.state->RemoveGroup(this);
-                        m_boundInstances[i].erase(itr++);
-                    }
-                    else
-                    {
-                        ++itr;
-                    }
-                }
-            }
-        }
-
-        // update the group's solo binds to the new leader
-        CharacterDatabase.PExecute("UPDATE `group_instance` SET `leaderGuid`='%u' WHERE `leaderGuid` = '%u'",
-                                   slot_lowguid, leader_lowguid);
-
-        // copy the permanent binds from the new leader to the group
-        // overwriting the solo binds with permanent ones if necessary
-        // in the DB those have been deleted already
-        Player::ConvertInstancesToGroup(player, this, slot->guid);
-
-        // update the group leader
-        CharacterDatabase.PExecute("UPDATE `groups` SET `leaderGuid`='%u' WHERE `groupId`='%u'", slot_lowguid, m_Id);
-        CharacterDatabase.CommitTransaction();
-    }
-
-    m_leaderGuid = slot->guid;
-    m_leaderName = slot->name;
-}
-
-/**
- * @brief Removes a player's participation from all active loot rolls.
- *
- * @param guid The player GUID to remove from roll tracking.
- */
-void Group::_removeRolls(ObjectGuid guid)
-{
-    for (Rolls::iterator it = RollId.begin(); it != RollId.end();)
-    {
-        Roll* roll = *it;
-        Roll::PlayerVote::iterator itr2 = roll->playerVote.find(guid);
-        if (itr2 == roll->playerVote.end())
-        {
-            ++it;
-            continue;
-        }
-
-        if (itr2->second == ROLL_GREED || itr2->second == ROLL_DISENCHANT)
-        {
-            --roll->totalGreed;
-        }
-        if (itr2->second == ROLL_NEED)
-        {
-            --roll->totalNeed;
-        }
-        if (itr2->second == ROLL_PASS)
-        {
-            --roll->totalPass;
-        }
-        if (itr2->second != ROLL_NOT_VALID)
-        {
-            --roll->totalPlayersRolling;
-        }
-
-        roll->playerVote.erase(itr2);
-
-        if (!CountRollVote(guid, it, ROLL_NOT_EMITED_YET))
-        {
-            ++it;
-        }
-    }
-}
-
-/**
- * @brief Changes a member's stored subgroup assignment.
- *
- * @param guid The member player GUID.
- * @param group The new subgroup.
- * @return true if the subgroup was updated; otherwise false.
- */
-bool Group::_setMembersGroup(ObjectGuid guid, uint8 group)
-{
-    member_witerator slot = _getMemberWSlot(guid);
-    if (slot == m_memberSlots.end())
-    {
-        return false;
-    }
-
-    slot->group = group;
-
-    SubGroupCounterIncrease(group);
-
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("UPDATE `group_member` SET `subgroup`='%u' WHERE `memberGuid`='%u'", group, guid.GetCounter());
-    }
-
-    return true;
-}
-
-/**
- * @brief Sets or clears the assistant flag for a member.
- *
- * @param guid The member player GUID.
- * @param state The desired assistant state.
- * @return true if the flag was updated; otherwise false.
- */
-bool Group::_setAssistantFlag(ObjectGuid guid, const bool& state)
-{
-    member_witerator slot = _getMemberWSlot(guid);
-    if (slot == m_memberSlots.end())
-    {
-        return false;
-    }
-
-    slot->assistant = state;
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("UPDATE `group_member` SET `assistant`='%u' WHERE `memberGuid`='%u'", (state == true) ? 1 : 0, guid.GetCounter());
-    }
-    return true;
-}
-
-/**
- * @brief Sets the group's main tank designation.
- *
- * @param guid The selected main tank GUID, or an empty GUID to clear it.
- * @return true if the designation changed; otherwise false.
- */
-bool Group::_setMainTank(ObjectGuid guid)
-{
-    if (m_mainTankGuid == guid)
-    {
-        return false;
-    }
-
-    if (guid)
-    {
-        member_citerator slot = _getMemberCSlot(guid);
-        if (slot == m_memberSlots.end())
-        {
-            return false;
-        }
-
-        if (m_mainAssistantGuid == guid)
-        {
-            _setMainAssistant(ObjectGuid());
-        }
-    }
-
-    m_mainTankGuid = guid;
-
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("UPDATE `groups` SET `mainTank`='%u' WHERE `groupId`='%u'", m_mainTankGuid.GetCounter(), m_Id);
-    }
-
-    return true;
-}
-
-/**
- * @brief Sets the group's main assistant designation.
- *
- * @param guid The selected main assistant GUID, or an empty GUID to clear it.
- * @return true if the designation changed; otherwise false.
- */
-bool Group::_setMainAssistant(ObjectGuid guid)
-{
-    if (m_mainAssistantGuid == guid)
-    {
-        return false;
-    }
-
-    if (guid)
-    {
-        member_witerator slot = _getMemberWSlot(guid);
-        if (slot == m_memberSlots.end())
-        {
-            return false;
-        }
-
-        if (m_mainTankGuid == guid)
-        {
-            _setMainTank(ObjectGuid());
-        }
-    }
-
-    m_mainAssistantGuid = guid;
-
-    if (!isBGGroup())
-        CharacterDatabase.PExecute("UPDATE `groups` SET `mainAssistant`='%u' WHERE `groupId`='%u'",
-                                   m_mainAssistantGuid.GetCounter(), m_Id);
-
-    return true;
-}
-
-/**
- * @brief Checks whether two players belong to the same subgroup of this group.
- *
- * @param member1 The first player.
- * @param member2 The second player.
- * @return true if both players belong to this group and share a subgroup; otherwise false.
- */
-bool Group::SameSubGroup(Player const* member1, Player const* member2) const
-{
-    if (!member1 || !member2)
-    {
-        return false;
-    }
-    if (member1->GetGroup() != this || member2->GetGroup() != this)
-    {
-        return false;
-    }
-    else
-    {
-        return member1->GetSubGroup() == member2->GetSubGroup();
-    }
-}
-
-// allows setting subgroup for offline members
-void Group::ChangeMembersGroup(ObjectGuid guid, uint8 group)
-{
-    if (!isRaidGroup())
-    {
-        return;
-    }
-
-    Player* player = sObjectMgr.GetPlayer(guid);
-
-    if (!player)
-    {
-        uint8 prevSubGroup = GetMemberGroup(guid);
-        if (prevSubGroup == group)
-        {
-            return;
-        }
-
-        if (_setMembersGroup(guid, group))
-        {
-            SubGroupCounterDecrease(prevSubGroup);
-            SendUpdate();
-        }
-    }
-    else
-        // This methods handles itself groupcounter decrease
-    {
-        ChangeMembersGroup(player, group);
-    }
-}
-
-// only for online members
-void Group::ChangeMembersGroup(Player* player, uint8 group)
-{
-    if (!player || !isRaidGroup())
-    {
-        return;
-    }
-
-    uint8 prevSubGroup = player->GetSubGroup();
-    if (prevSubGroup == group)
-    {
-        return;
-    }
-
-    if (_setMembersGroup(player->GetObjectGuid(), group))
-    {
-        if (player->GetGroup() == this)
-        {
-            player->GetGroupRef().setSubGroup(group);
-        }
-        // if player is in BG raid, it is possible that he is also in normal raid - and that normal raid is stored in m_originalGroup reference
-        else
-        {
-            prevSubGroup = player->GetOriginalSubGroup();
-            player->GetOriginalGroupRef().setSubGroup(group);
-        }
-        SubGroupCounterDecrease(prevSubGroup);
-
-        SendUpdate();
-    }
-}
-
-uint32 Group::GetMaxSkillValueForGroup(SkillType skill)
-{
-    uint32 maxvalue = 0;
-
-    for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-    {
-        Player* member = itr->getSource();
-        if (!member)
-        {
-            continue;
-        }
-
-        uint32 value = member->GetSkillValue(skill);
-        if (maxvalue < value)
-        {
-            maxvalue = value;
-        }
-    }
-
-    return maxvalue;
-}
-
-/**
- * @brief Advances the round-robin looter to the next eligible nearby member.
- *
- * @param pSource The looted world object.
- * @param ifneed True to keep the current looter when still eligible.
- */
-void Group::UpdateLooterGuid(WorldObject* pSource, bool ifneed)
-{
-    switch (GetLootMethod())
-    {
-        case MASTER_LOOT:
-        case FREE_FOR_ALL:
-            return;
-        default:
-            // round robin style looting applies for all low
-            // quality items in each loot method except free for all and master loot
-            break;
-    }
-
-    member_citerator guid_itr = _getMemberCSlot(GetLooterGuid());
-    if (guid_itr != m_memberSlots.end())
-    {
-        if (ifneed)
-        {
-            // not update if only update if need and ok
-            Player* looter = sObjectAccessor.FindPlayer(guid_itr->guid);
-            if (looter && looter->IsWithinDist(pSource, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE), false))
-            {
-                return;
-            }
-        }
-        ++guid_itr;
-    }
-
-    // search next after current
-    if (guid_itr != m_memberSlots.end())
-    {
-        for (member_citerator itr = guid_itr; itr != m_memberSlots.end(); ++itr)
-        {
-            if (Player* pl = sObjectAccessor.FindPlayer(itr->guid))
-            {
-                if (pl->IsWithinDist(pSource, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE), false))
-                {
-                    bool refresh = pl->GetLootGuid() == pSource->GetObjectGuid();
-
-                    // if (refresh)                          // update loot for new looter
-                    //    pl->GetSession()->DoLootRelease(pl->GetLootGUID());
-                    SetLooterGuid(pl->GetObjectGuid());
-                    SendUpdate();
-                    if (refresh)                            // update loot for new looter
-                    {
-                        pl->SendLoot(pSource->GetObjectGuid(), LOOT_CORPSE);
-                    }
-                    return;
-                }
-            }
-        }
-    }
-
-    // search from start
-    for (member_citerator itr = m_memberSlots.begin(); itr != guid_itr; ++itr)
-    {
-        if (Player* pl = sObjectAccessor.FindPlayer(itr->guid))
-        {
-            if (pl->IsWithinDist(pSource, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE), false))
-            {
-                bool refresh = pl->GetLootGuid() == pSource->GetObjectGuid();
-
-                // if (refresh)                              // update loot for new looter
-                //    pl->GetSession()->DoLootRelease(pl->GetLootGUID());
-                SetLooterGuid(pl->GetObjectGuid());
-                SendUpdate();
-                if (refresh)                                // update loot for new looter
-                {
-                    pl->SendLoot(pSource->GetObjectGuid(), LOOT_CORPSE);
-                }
-                return;
-            }
-        }
-    }
-
-    SetLooterGuid(ObjectGuid());
-    SendUpdate();
-}
-
-/**
- * @brief Validates whether the full group can join a battleground queue together.
- *
- * @param bgTypeId The battleground type identifier.
- * @param bgQueueTypeId The battleground queue type identifier.
- * @param MinPlayerCount The minimum allowed group size.
- * @param MaxPlayerCount The maximum allowed group size.
- * @return uint32 A battleground join error code.
- */
-GroupJoinBattlegroundResult Group::CanJoinBattleGroundQueue(BattleGround const* bgOrTemplate, BattleGroundQueueTypeId bgQueueTypeId, uint32 MinPlayerCount, uint32 MaxPlayerCount, bool isRated, uint32 arenaSlot)
-{
-    BattlemasterListEntry const* bgEntry = sBattlemasterListStore.LookupEntry(bgOrTemplate->GetTypeID());
-    if (!bgEntry)
-    {
-        return ERR_GROUP_JOIN_BATTLEGROUND_FAIL;            // shouldn't happen
-    }
-
-    // check for min / max count
-    uint32 memberscount = GetMembersCount();
-
-    // only check for MinPlayerCount since MinPlayerCount == MaxPlayerCount for arenas...
-    if (bgOrTemplate->isArena() && memberscount != MinPlayerCount)
-    {
-        return ERR_ARENA_TEAM_PARTY_SIZE;
-    }
-
-    if (memberscount > bgEntry->maxGroupSize)               // no MinPlayerCount for battlegrounds
-    {
-        return ERR_BATTLEGROUND_NONE;                       // ERR_GROUP_JOIN_BATTLEGROUND_TOO_MANY handled on client side
-    }
-
-    // get a player as reference, to compare other players' stats to (arena team id, queue id based on level, etc.)
-    Player* reference = GetFirstMember()->getSource();
-    // no reference found, can't join this way
-    if (!reference)
-    {
-        return ERR_BATTLEGROUND_JOIN_FAILED;
-    }
-
-    PvPDifficultyEntry const* bracketEntry = GetBattlegroundBracketByLevel(bgOrTemplate->GetMapId(), reference->getLevel());
-    if (!bracketEntry)
-    {
-        return ERR_BATTLEGROUND_JOIN_FAILED;
-    }
-
-    uint32 arenaTeamId = reference->GetArenaTeamId(arenaSlot);
-    Team team = reference->GetTeam();
-
-    BattleGroundQueueTypeId bgQueueTypeIdRandom = BattleGroundMgr::BGQueueTypeId(BATTLEGROUND_RB, ARENA_TYPE_NONE);
-
-    // check every member of the group to be able to join
-    for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-    {
-        Player* member = itr->getSource();
-        // offline member? don't let join
-        if (!member)
-        {
-            return ERR_BATTLEGROUND_JOIN_FAILED;
-        }
-        // don't allow cross-faction join as group
-        if (member->GetTeam() != team)
-        {
-            return ERR_BATTLEGROUND_JOIN_TIMED_OUT;
-        }
-        // not in the same battleground level bracket, don't let join
-        PvPDifficultyEntry const* memberBracketEntry = GetBattlegroundBracketByLevel(bracketEntry->mapId, member->getLevel());
-        if (memberBracketEntry != bracketEntry)
-        {
-            return ERR_BATTLEGROUND_JOIN_RANGE_INDEX;
-        }
-        // don't let join rated matches if the arena team id doesn't match
-        if (isRated && member->GetArenaTeamId(arenaSlot) != arenaTeamId)
-        {
-            return ERR_BATTLEGROUND_JOIN_FAILED;
-        }
-        // don't let join if someone from the group is already in that bg queue
-        if (member->InBattleGroundQueueForBattleGroundQueueType(bgQueueTypeId))
-        {
-            return ERR_BATTLEGROUND_JOIN_FAILED;            // not blizz-like
-        }
-        // don't let join if someone from the group is in bg queue random
-        if (member->InBattleGroundQueueForBattleGroundQueueType(bgQueueTypeIdRandom))
-        {
-            return ERR_IN_RANDOM_BG;
-        }
-        // don't let join to bg queue random if someone from the group is already in bg queue
-        if (bgOrTemplate->GetTypeID() == BATTLEGROUND_RB && member->InBattleGroundQueue())
-        {
-            return ERR_IN_NON_RANDOM_BG;
-        }
-        // check for deserter debuff in case not arena queue
-        if (bgOrTemplate->GetTypeID() != BATTLEGROUND_AA && !member->CanJoinToBattleground())
-        {
-            return ERR_GROUP_JOIN_BATTLEGROUND_DESERTERS;
-        }
-        // check if member can join any more battleground queues
-        if (!member->HasFreeBattleGroundQueueId())
-        {
-            return ERR_BATTLEGROUND_TOO_MANY_QUEUES;        // not blizz-like
-        }
-    }
-    return GroupJoinBattlegroundResult(bgOrTemplate->GetTypeID());
-}
-
-void Group::SetDungeonDifficulty(Difficulty difficulty)
-{
-    m_dungeonDifficulty = difficulty;
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("UPDATE `groups` SET `difficulty` = %u WHERE `groupId`='%u'", m_dungeonDifficulty, m_Id);
-    }
-
-    for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-    {
-        Player* player = itr->getSource();
-        if (!player->GetSession() || player->getLevel() < LEVELREQUIREMENT_HEROIC)
-        {
-            continue;
-        }
-        player->SetDungeonDifficulty(difficulty);
-        player->SendDungeonDifficulty(true);
-    }
-}
-
-void Group::SetRaidDifficulty(Difficulty difficulty)
-{
-    m_raidDifficulty = difficulty;
-    if (!isBGGroup())
-    {
-        CharacterDatabase.PExecute("UPDATE `groups` SET `raiddifficulty` = %u WHERE `groupId` = '%u'", m_raidDifficulty, m_Id);
-    }
-
-    for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-    {
-        Player* player = itr->getSource();
-        if (!player->GetSession() || player->getLevel() < LEVELREQUIREMENT_HEROIC)
-        {
-            continue;
-        }
-        player->SetRaidDifficulty(difficulty);
-        player->SendRaidDifficulty(true);
-    }
-}
-
-/**
- * @brief Checks whether any group member is in combat inside a specific instance.
- *
- * @param instanceId The instance identifier to test.
- * @return true if a member in that instance currently has attackers; otherwise false.
- */
-bool Group::InCombatToInstance(uint32 instanceId)
-{
-    for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-    {
-        Player* pPlayer = itr->getSource();
-        if (pPlayer->getAttackers().size() && pPlayer->GetInstanceId() == instanceId)
-        {
-            return true;
-        }
-    }
-    return false;
-}
-
-bool Group::SetPlayerMap(ObjectGuid guid, uint32 mapid)
-{
-    member_witerator slot = _getMemberWSlot(guid);
-    if (slot != m_memberSlots.end())
-    {
-        slot->lastMap = mapid;
-        DEBUG_LOG("Group::SetPlayerMap> map is updated");
-        return true;
-    }
-    return false;
-}
-
-/**
- * @brief Resets or unbinds group dungeon instances according to the requested method.
- *
- * @param method The reset method to apply.
- * @param SendMsgTo Optional player who should receive reset result messages.
- */
-void Group::ResetInstances(InstanceResetMethod method, bool isRaid, Player* SendMsgTo)
-{
-    if (isBGGroup())
-    {
-        return;
-    }
-
-    // method can be INSTANCE_RESET_ALL, INSTANCE_RESET_CHANGE_DIFFICULTY, INSTANCE_RESET_GROUP_DISBAND
-
-    // we assume that when the difficulty changes, all instances that can be reset will be
-    Difficulty diff = GetDifficulty(isRaid);
-
-    typedef std::set<uint32> OfflineMapSet;
-    OfflineMapSet mapsWithOfflinePlayer;                    // to store map of offline players
-
-    if (method != INSTANCE_RESET_GROUP_DISBAND)
-    {
-        // Store maps in which are offline members for instance reset check.
-        for (member_citerator itr = m_memberSlots.begin(); itr != m_memberSlots.end(); ++itr)
-        {
-            if (!sObjectAccessor.FindPlayer(itr->guid))
-            {
-                mapsWithOfflinePlayer.insert(itr->lastMap); // add last map from offline player
-            }
-        }
-    }
-
-    for (BoundInstancesMap::iterator itr = m_boundInstances[diff].begin(); itr != m_boundInstances[diff].end();)
-    {
-        DungeonPersistentState* state = itr->second.state;
-        const MapEntry* entry = sMapStore.LookupEntry(itr->first);
-        if (!entry || entry->IsRaid() != isRaid || (!state->CanReset() && method != INSTANCE_RESET_GROUP_DISBAND))
-        {
-            ++itr;
-            continue;
-        }
-
-        if (method == INSTANCE_RESET_ALL)
-        {
-            // the "reset all instances" method can only reset normal maps
-            if (entry->map_type == MAP_RAID || diff == DUNGEON_DIFFICULTY_HEROIC)
-            {
-                ++itr;
-                continue;
-            }
-        }
-
-        bool isEmpty = true;
-        // check if there are offline members on the map
-        if (method != INSTANCE_RESET_GROUP_DISBAND && mapsWithOfflinePlayer.find(state->GetMapId()) != mapsWithOfflinePlayer.end())
-        {
-            isEmpty = false;
-        }
-
-        // if the map is loaded, reset it if can
-        if (isEmpty && entry->IsDungeon() && !(method == INSTANCE_RESET_GROUP_DISBAND && !state->CanReset()))
-            if (Map* map = sMapMgr.FindMap(state->GetMapId(), state->GetInstanceId()))
-            {
-                isEmpty = ((DungeonMap*)map)->Reset(method);
-            }
-
-        if (SendMsgTo)
-        {
-            if (isEmpty)
-            {
-                SendMsgTo->SendResetInstanceSuccess(state->GetMapId());
-            }
-            else
-            {
-                SendMsgTo->SendResetInstanceFailed(0, state->GetMapId());
-            }
-        }
-
-        // TODO - Adapt here when clear how difficulty changes must be handled
-        if (isEmpty || method == INSTANCE_RESET_GROUP_DISBAND || method == INSTANCE_RESET_CHANGE_DIFFICULTY)
-        {
-            // do not reset the instance, just unbind if others are permanently bound to it
-            if (state->CanReset())
-            {
-                state->DeleteFromDB();
-            }
-            else
-            {
-                CharacterDatabase.PExecute("DELETE FROM `group_instance` WHERE `instance` = '%u'", state->GetInstanceId());
-            }
-            // i don't know for sure if hash_map iterators
-            m_boundInstances[diff].erase(itr);
-            itr = m_boundInstances[diff].begin();
-            // this unloads the instance save unless online players are bound to it
-            // (eg. permanent binds or GM solo binds)
-            state->RemoveGroup(this);
-        }
-        else
-        {
-            ++itr;
-        }
-    }
-}
-
-/**
- * @brief Retrieves the group's binding information for a dungeon map.
- *
- * @param mapid The map identifier.
- * @return InstanceGroupBind* The bind information if present; otherwise NULL.
- */
-InstanceGroupBind* Group::GetBoundInstance(uint32 mapid, Player* player)
-{
-    MapEntry const* mapEntry = sMapStore.LookupEntry(mapid);
-    if (!mapEntry)
-    {
-        return NULL;
-    }
-
-    Difficulty difficulty = player->GetDifficulty(mapEntry->IsRaid());
-
-    // some instances only have one difficulty
-    MapDifficultyEntry const* mapDiff = GetMapDifficultyData(mapid, difficulty);
-    if (!mapDiff)
-    {
-        difficulty = DUNGEON_DIFFICULTY_NORMAL;
-    }
-
-    BoundInstancesMap::iterator itr = m_boundInstances[difficulty].find(mapid);
-    if (itr != m_boundInstances[difficulty].end())
-    {
-        return &itr->second;
-    }
-    else
-    {
-        return NULL;
-    }
-}
-
-InstanceGroupBind* Group::GetBoundInstance(Map* aMap, Difficulty difficulty)
-{
-    // some instances only have one difficulty
-    MapDifficultyEntry const* mapDiff = GetMapDifficultyData(aMap->GetId(), difficulty);
-    if (!mapDiff)
-    {
-        return NULL;
-    }
-
-    BoundInstancesMap::iterator itr = m_boundInstances[difficulty].find(aMap->GetId());
-    if (itr != m_boundInstances[difficulty].end())
-    {
-        return &itr->second;
-    }
-    else
-    {
-        return NULL;
-    }
-}
-
-/**
- * @brief Binds the group to a dungeon instance persistent state.
- *
- * @param state The dungeon persistent state to bind.
- * @param permanent True to create a permanent bind.
- * @param load True when reconstructing binds from storage.
- * @return InstanceGroupBind* The updated bind information, or NULL on failure.
- */
-InstanceGroupBind* Group::BindToInstance(DungeonPersistentState* state, bool permanent, bool load)
-{
-    if (state && !isBGGroup())
-    {
-        InstanceGroupBind& bind = m_boundInstances[state->GetDifficulty()][state->GetMapId()];
-        if (bind.state)
-        {
-            // when a boss is killed or when copying the players's binds to the group
-            if (permanent != bind.perm || state != bind.state)
-                if (!load)
-                    CharacterDatabase.PExecute("UPDATE `group_instance` SET `instance` = '%u', `permanent` = '%u' WHERE `leaderGuid` = '%u' AND `instance` = '%u'",
-                                               state->GetInstanceId(), permanent, GetLeaderGuid().GetCounter(), bind.state->GetInstanceId());
-        }
-        else if (!load)
-            CharacterDatabase.PExecute("INSERT INTO `group_instance` (`leaderGuid`, `instance`, `permanent`) VALUES ('%u', '%u', '%u')",
-                                       GetLeaderGuid().GetCounter(), state->GetInstanceId(), permanent);
-
-        if (bind.state != state)
-        {
-            if (bind.state)
-            {
-                bind.state->RemoveGroup(this);
-            }
-            state->AddGroup(this);
-        }
-
-        bind.state = state;
-        bind.perm = permanent;
-        if (!load)
-            DEBUG_LOG("Group::BindToInstance: Group (Id: %d) is now bound to map %d, instance %d, difficulty %d",
-                      GetId(), state->GetMapId(), state->GetInstanceId(), state->GetDifficulty());
-        return &bind;
-    }
-    else
-    {
-        return NULL;
-    }
-}
-
-/**
- * @brief Removes the group's binding to a dungeon instance.
- *
- * @param mapid The map identifier to unbind.
- * @param unload True when the unbind happens as part of loading or teardown logic.
- */
-void Group::UnbindInstance(uint32 mapid, uint8 difficulty, bool unload)
-{
-    BoundInstancesMap::iterator itr = m_boundInstances[difficulty].find(mapid);
-    if (itr != m_boundInstances[difficulty].end())
-    {
-        if (!unload)
-            CharacterDatabase.PExecute("DELETE FROM `group_instance` WHERE `leaderGuid` = '%u' AND `instance` = '%u'",
-                                       GetLeaderGuid().GetCounter(), itr->second.state->GetInstanceId());
-        itr->second.state->RemoveGroup(this);               // save can become invalid
-        m_boundInstances[difficulty].erase(itr);
-    }
-}
-
-/**
- * @brief Marks a player as invalid for their current instance when group removal requires homebinding.
- *
- * @param player The player leaving the group.
- */
-void Group::_homebindIfInstance(Player* player)
-{
-    if (player && !player->isGameMaster())
-    {
-        Map* map = player->GetMap();
-        if (map->IsDungeon())
-        {
-            // leaving the group in an instance, the homebind timer is started
-            // unless the player is permanently saved to the instance
-            InstancePlayerBind* playerBind = player->GetBoundInstance(map->GetId(), map->GetDifficulty());
-            if (!playerBind || !playerBind->perm)
-            {
-                player->m_InstanceValid = false;
-            }
-        }
-    }
-}
-
-/**
- * @brief Awards honor, reputation, XP, pet XP, and kill credit to one group member.
- *
- * @param pGroupGuy The group member receiving rewards.
- * @param pVictim The defeated unit.
- * @param count The number of qualifying group members.
- * @param PvP True when the kill is treated as PvP.
- * @param group_rate The group XP scaling factor.
- * @param sum_level The summed levels of qualifying members.
- * @param is_dungeon True when the kill occurred in a dungeon.
- * @param not_gray_member_with_max_level The highest-level member still eligible for XP.
- * @param member_with_max_level The highest-level qualifying group member.
- * @param xp The base XP amount for the kill.
- */
-static void RewardGroupAtKill_helper(Player* pGroupGuy, Unit* pVictim, uint32 count, bool PvP, float group_rate, uint32 sum_level, bool is_dungeon, Player* not_gray_member_with_max_level, Player* member_with_max_level, uint32 xp)
-{
-    // honor can be in PvP and !PvP (racial leader) cases (for alive)
-    if (pGroupGuy->IsAlive())
-    {
-        pGroupGuy->RewardHonor(pVictim, count);
-    }
-
-    // xp and reputation only in !PvP case
-    if (!PvP)
-    {
-        float rate = group_rate * float(pGroupGuy->getLevel()) / sum_level;
-
-        // if is in dungeon then all receive full reputation at kill
-        // rewarded any alive/dead/near_corpse group member
-        pGroupGuy->RewardReputation(pVictim, is_dungeon ? 1.0f : rate);
-
-        // XP updated only for alive group member
-        if (pGroupGuy->IsAlive() && not_gray_member_with_max_level &&
-            pGroupGuy->getLevel() <= not_gray_member_with_max_level->getLevel())
-        {
-            uint32 itr_xp = (member_with_max_level == not_gray_member_with_max_level) ? uint32(xp * rate) : uint32((xp * rate / 2) + 1);
-
-            pGroupGuy->GiveXP(itr_xp, pVictim);
-            if (Pet* pet = pGroupGuy->GetPet())
-            {
-                pet->GivePetXP(itr_xp / 2);
-            }
-        }
-
-        // quest objectives updated only for alive group member or dead but with not released body
-        if (pGroupGuy->IsAlive() || !pGroupGuy->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
-        {
-            // normal creature (not pet/etc) can be only in !PvP case
-            if (pVictim->GetTypeId() == TYPEID_UNIT)
-                if (CreatureInfo const* normalInfo = ObjectMgr::GetCreatureTemplate(pVictim->GetEntry()))
-                {
-                    pGroupGuy->KilledMonster(normalInfo, pVictim->GetObjectGuid());
-                }
-        }
-    }
-}
-
-/** Provide rewards to group members at unit kill
- *
- * @param pVictim       Killed unit
- * @param player_tap    Player who tap unit if online, it can be group member or can be not if leaved after tap but before kill target
- *
- * Rewards received by group members and player_tap
- */
-void Group::RewardGroupAtKill(Unit* pVictim, Player* player_tap)
-{
-    bool PvP = pVictim->IsCharmedOwnedByPlayerOrPlayer();
-
-    // prepare data for near group iteration (PvP and !PvP cases)
-    uint32 xp = 0;
-
-    uint32 count = 0;
-    uint32 sum_level = 0;
-    Player* member_with_max_level = NULL;
-    Player* not_gray_member_with_max_level = NULL;
-
-    GetDataForXPAtKill(pVictim, count, sum_level, member_with_max_level, not_gray_member_with_max_level, player_tap);
-
-    if (member_with_max_level)
-    {
-        /// not get Xp in PvP or no not gray players in group
-        xp = (PvP || !not_gray_member_with_max_level) ? 0 : MaNGOS::XP::Gain(not_gray_member_with_max_level, pVictim);
-
-        /// skip in check PvP case (for speed, not used)
-        bool is_raid = PvP ? false : sMapStore.LookupEntry(pVictim->GetMapId())->IsRaid() && isRaidGroup();
-        bool is_dungeon = PvP ? false : sMapStore.LookupEntry(pVictim->GetMapId())->IsDungeon();
-        float group_rate = MaNGOS::XP::xp_in_group_rate(count, is_raid);
-
-        for (GroupReference* itr = GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* pGroupGuy = itr->getSource();
-            if (!pGroupGuy)
-            {
-                continue;
-            }
-
-            // will proccessed later
-            if (pGroupGuy == player_tap)
-            {
-                continue;
-            }
-
-            if (!pGroupGuy->IsAtGroupRewardDistance(pVictim))
-            {
-                continue;                                // member (alive or dead) or his corpse at req. distance
-            }
-
-            RewardGroupAtKill_helper(pGroupGuy, pVictim, count, PvP, group_rate, sum_level, is_dungeon, not_gray_member_with_max_level, member_with_max_level, xp);
-        }
-
-        if (player_tap)
-        {
-            // member (alive or dead) or his corpse at req. distance
-            if (player_tap->IsAtGroupRewardDistance(pVictim))
-            {
-                RewardGroupAtKill_helper(player_tap, pVictim, count, PvP, group_rate, sum_level, is_dungeon, not_gray_member_with_max_level, member_with_max_level, xp);
-            }
-        }
-    }
-}
